use docketNumber instead of caseId for case deadlines

+++ shared/src/persistence/dynamo/caseDeadlines/deleteCaseDeadline.js
@@ -1,4 +1,7 @@
 const client = require('../../dynamodbClientService');
+const {
+  getCaseIdFromDocketNumber,
+} = require('../cases/getCaseIdFromDocketNumber');
 
 /**
  * deleteCaseDeadline
@@ -6,45 +9,52 @@ const client = require('../../dynamodbClientService');
  * @param {object} providers the providers object
  * @param {object} providers.applicationContext the application context
  * @param {string} providers.caseDeadlineId the id of the case deadline to delete
- * @param {string} providers.caseId the id of the case the deadline is attached to
+ * @param {string} providers.docketNumber the docket number of the case the deadline is attached to
  * @returns {Array<Promise>} the promises for the persistence delete calls
  */
 exports.deleteCaseDeadline = async ({
   applicationContext,
   caseDeadlineId,
-  caseId,
+  docketNumber,
 }) => {
-  const results = [];
+  const caseId = await getCaseIdFromDocketNumber({
+    applicationContext,
+    docketNumber,
+  });
 
-  results.push(
-    await client.delete({
-      applicationContext,
-      key: {
-        pk: `case-deadline|${caseDeadlineId}`,
-        sk: `case-deadline|${caseDeadlineId}`,
-      },
-    }),
-  );
+  if (caseId) {
+    const results = [];
 
-  results.push(
-    await client.delete({
-      applicationContext,
-      key: {
-        pk: `case|${caseId}`,
-        sk: `case-deadline|${caseDeadlineId}`,
-      },
-    }),
-  );
+    results.push(
+      await client.delete({
+        applicationContext,
+        key: {
+          pk: `case-deadline|${caseDeadlineId}`,
+          sk: `case-deadline|${caseDeadlineId}`,
+        },
+      }),
+    );
 
-  results.push(
-    await client.delete({
-      applicationContext,
-      key: {
-        pk: 'case-deadline-catalog',
-        sk: `case-deadline|${caseDeadlineId}`,
-      },
-    }),
-  );
+    results.push(
+      await client.delete({
+        applicationContext,
+        key: {
+          pk: `case|${caseId}`,
+          sk: `case-deadline|${caseDeadlineId}`,
+        },
+      }),
+    );
 
-  return results;
+    results.push(
+      await client.delete({
+        applicationContext,
+        key: {
+          pk: 'case-deadline-catalog',
+          sk: `case-deadline|${caseDeadlineId}`,
+        },
+      }),
+    );
+
+    return results;
+  }
 };
